+++ docs/ci-cd.md
@@ -2,11 +2,91 @@
 
 While it is conveninet to leveage the CDK command line tool to deploy your first cluster, we reccomend setting up automated pipelines that will be responsible for deploying and updating your EKS infrastructure. 
 
-To accomplish this, the EKS SSP - Refernce Solution leverages the [`CDK Pipelines`](https://docs.aws.amazon.com/cdk/api/latest/docs/pipelines-readme.html) module. This module makes it trivial to create Continuous Delivery pipelines via CodePipeline that are responsible for deploying and updating your infrastructure. 
+To accomplish this, the EKS SSP - Refernce Solution leverages the [`Pipelines`](https://docs.aws.amazon.com/cdk/api/latest/docs/pipelines-readme.html) CDK module. This module makes it trivial to create Continuous Delivery (CD) pipelines via CodePipeline that are responsible for deploying and updating your infrastructure. 
 
-Aditionally, the the EKS SSP - Refernce Solution leverages the GitHub integration that the `CDK Pipelines` module provides in order to integrate our pipelines with Git. The end result is that any new configuration pushed to a GitHub repository containing our CDK will be automatically deployed.
+Aditionally, the the EKS SSP - Refernce Solution leverages the GitHub integration that the `Pipelines` CDK module provides in order to integrate our pipelines with Git. The end result is that any new configuration pushed to a GitHub repository containing our CDK will be automatically deployed.
 
-## Building our Pipeline
+## Complete Implementation
+
+The below code example provides a complete example for how to leverag the `Pipelines` CDK module to setup a CD pipeline for an EKS blueprint. Below, we will break down each code section individually. 
+
+```typescript
+import * as cdk from '@aws-cdk/core';
+import * as pipelines from '@aws-cdk/pipelines';
+import * as codepipeline from '@aws-cdk/aws-codepipeline';
+import * as actions from '@aws-cdk/aws-codepipeline-actions';
+
+// SSP Lib
+import * as ssp from '../../lib'
+
+// Team implementations
+import * as team from '../teams'
+
+class Pipeline {
+    public static build = (scope: cdk.App) => {
+        // Github action.
+        const sourceArtifact = new codepipeline.Artifact();
+        const oathToken = cdk.SecretValue.secretsManager('github-token')
+        const sourceAction = new actions.GitHubSourceAction({
+            actionName: `pipeline-github-action`,
+            owner: 'aws-quickstarts',
+            repo: 'quickstart-ssp-amazon-eks',
+            branch: 'main',
+            output: sourceArtifact,
+            oauthToken: oathToken,
+        })
+
+        // Synth action.
+        const cloudAssemblyArtifact = new codepipeline.Artifact();
+        const synthAction = pipelines.SimpleSynthAction.standardNpmSynth({
+            sourceArtifact,
+            cloudAssemblyArtifact,
+            buildCommand: 'npm run build',
+        })
+
+        return new pipelines.CdkPipeline(scope, 'blueprint-pipeline', {
+            pipelineName: 'blueprint-pipeline',
+            cloudAssemblyArtifact,
+            sourceAction,
+            synthAction
+        });
+    }
+}
+
+class ClusterStage extends cdk.Stage {
+    constructor(scope: cdk.Stack, id: string, props?: cdk.StageProps) {
+        super(scope, id, props);
+
+        // Setup platform team
+        const accountID = props?.env?.account
+        const platformTeam = new team.TeamPlatform(accountID!)
+        const teams: Array<ssp.Team> = [platformTeam];
+
+        // AddOns for the cluster.
+        const addOns: Array<ssp.ClusterAddOn> = [
+            new ssp.NginxAddOn,
+            new ssp.ArgoCDAddOn,
+            new ssp.CalicoAddOn,
+            new ssp.MetricsServerAddOn,
+            new ssp.ClusterAutoScalerAddOn,
+            new ssp.ContainerInsightsAddOn,
+        ];
+        new ssp.EksBlueprint(this, { id: 'eks', addOns, teams }, props);
+    }
+}
+
+class PipelineStack extends cdk.Stack {
+    constructor(scope: cdk.App, id: string, props?: cdk.StackProps) {
+        super(scope, id)
+
+        const pipeline = Pipeline.build(scope)
+        const dev = new ClusterStage(this, 'blueprint-cluster-stage')
+        pipeline.addApplicationStage(dev);
+    }
+}
+```
+
+## Breaking it down. 
 
 ### Create a GitHub Source Action
